+++ packages/ricos-content/web/src/converters/draft/fromDraft/__tests__/fromDraft.spec.ts
@@ -15,6 +15,8 @@ import noVideoData from './no-data-video.json';
 import noVideoDataMigrated from './no-data-video-migrated.json';
 import innerWithUnsupported from './inner-rce-with-unsupported-styling.json';
 import innerWithUnsupportedMigrated from './inner-rce-with-unsupported-styling-migrated.json';
+import faultyLinkValues from './faulty-link-values.json';
+import faultyLinkValuesMigrated from './faulty-link-values-migrated.json';
 import polyfills from '../../../../../../../../e2e/tests/fixtures/polyfills.json';
 import { getTextNodes } from '../getTextNodes';
 import complexRicosFixture from '../../../../../statics/json/migratedFixtures/migration-content.json';
@@ -374,13 +376,24 @@ describe('migrate from draft', () => {
     ).toEqual({});
   });
 
-  it('should convert faulty content correctly', () => {
+  it('should convert unsupported inine inner content correctly', () => {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     const converted = fromDraft(innerWithUnsupported, { ignoreUnsupportedValues: true });
 
     expect(
       compare(converted, RichContent.fromJSON(innerWithUnsupportedMigrated), {
+        ignoredKeys: ['id'],
+      })
+    ).toEqual({});
+  });
+
+  it('should convert faulty link content correctly', () => {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const converted = fromDraft(faultyLinkValues, { ignoreUnsupportedValues: true });
+    expect(
+      compare(converted, RichContent.fromJSON(faultyLinkValuesMigrated), {
         ignoredKeys: ['id'],
       })
     ).toEqual({});